refactor: move the product version reader to a shared helper
